First version of ooba-powered setup_local_text_llm
