+++ pyproject.toml
@@ -24,11 +24,12 @@ python-dotenv = "^1.0.0"
 # On windows, `pyreadline3` replaces that, so you can also just `import readline`.
 inquirer = "^3.1.3"
 wget = "^3.2"
-huggingface-hub = "^0.16.4"
+huggingface-hub = "^0.17.3"
 litellm = "^0.1.590"
 pyyaml = "^6.0.1"
 semgrep = "^1.41.0"
 yaspin = "^3.0.1"
+ooba = "^0.0.16"
 [tool.poetry.dependencies.pyreadline3]
 version = "^3.4.1"
 markers = "sys_platform == 'win32'"
